refactor changes using current hooks
refactored changes
